Add support for OBS Mirror OpenXR layer
Committed in as a binary because we're loading quad-views-foveated
explicitly, and there's no way to change the load order of OpenXR
layers. They always go implicit first, then explicit.

In order for the mirror layer to work with foveated rendering,
quad-views-foveated needs to first take in the layer with 4 views and do
the magic to merge them into a stereo view. The mirror layer is then
able to display the merged views properly.